Split window module into its own `sfml-window` crate
