+++ sfml-graphics/examples/custom-drawable.rs
@@ -1,9 +1,10 @@
 extern crate sfml;
+extern crate sfml_window;
 extern crate sfml_graphics;
 
 use sfml_graphics::{CircleShape, Color, Drawable, RectangleShape, RenderStates, RenderTarget,
                     RenderWindow, Shape, Transformable};
-use sfml::window::{Key, VideoMode, Event, style};
+use sfml_window::{Key, VideoMode, Event, style};
 use sfml::system::Vector2f;
 
 /// Our custom drawable type. It looks like a bullet.
