+++ sfml-window/src/style.rs
